Reorganize code distribution in files; add comments + test
