+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "fido-authenticator"
 version = "0.0.0-unreleased"
 authors = ["Nicolas Stalder <n@stalder.io>"]
-edition = "2018"
+edition = "2021"
 license = "Apache-2.0 OR MIT"
 repository = "https://github.com/solokeys/fido-authenticator"
 documentation = "https://docs.rs/fido-authenticator"
@@ -22,9 +22,8 @@ ctap-types = { git = "https://github.com/solokeys/ctap-types" }
 trussed = { git = "https://github.com/trussed-dev/trussed" }
 
 [features]
-enable-fido-pre = []
-
 disable-reset-time-window = []
+enable-fido-pre = []
 
 log-all = []
 log-none = []
@@ -32,3 +31,7 @@ log-info = []
 log-debug = []
 log-warn = []
 log-error = []
+
+# [dev-dependencies]
+# quickcheck = "1"
+# rand = "0.8.4"
